Skip updating the version if no changes have been found
The non-Plain-Version Version string that we generate contains
the current timestamp, so any run of the script would generate
a diff in the Version: line, triggering needless rebuilds and
source checkins. Only change the Version line if there was a
code change. In cases when the %setup line or the source tarball
does not match, still do the substitution of those.